Extract Paginator to use across all services
