+++ moto/logs/models.py
@@ -2,12 +2,14 @@ from boto3 import Session
 
 from moto.core import BaseBackend
 from moto.core.utils import unix_time_millis
+from moto.utilities.Paginator import paginate
 from .exceptions import (
     ResourceNotFoundException,
     ResourceAlreadyExistsException,
     InvalidParameterException,
     LimitExceededException,
 )
+from .utils import PAGINATION_MODEL
 
 
 class LogEvent:
@@ -265,14 +267,15 @@ class LogGroup:
             raise ResourceNotFoundException()
         del self.streams[log_stream_name]
 
+    @paginate(pagination_model=PAGINATION_MODEL)
     def describe_log_streams(
         self,
         descending,
-        limit,
         log_group_name,
         log_stream_name_prefix,
-        next_token,
         order_by,
+        next_token=None,
+        limit=None,
     ):
         # responses only logStreamName, creationTime, arn, storedBytes when no events are stored.
 
@@ -289,16 +292,9 @@ class LogGroup:
                 else item[1].get("lastEventTimestamp", 0)
             )
 
-        if next_token is None:
-            next_token = 0
-
         log_streams = sorted(log_streams, key=sorter, reverse=descending)
-        new_token = next_token + limit
-        log_streams_page = [x[1] for x in log_streams[next_token:new_token]]
-        if new_token >= len(log_streams):
-            new_token = None
-
-        return log_streams_page, new_token
+        log_streams = [item[1] for item in log_streams]
+        return log_streams
 
     def put_log_events(
         self, log_group_name, log_stream_name, log_events, sequence_token
@@ -482,7 +478,8 @@ class LogsBackend(BaseBackend):
             raise ResourceNotFoundException()
         del self.groups[log_group_name]
 
-    def describe_log_groups(self, limit, log_group_name_prefix, next_token):
+    @paginate(pagination_model=PAGINATION_MODEL)
+    def describe_log_groups(self, log_group_name_prefix, limit=None, next_token=None):
         if log_group_name_prefix is None:
             log_group_name_prefix = ""
 
@@ -493,33 +490,7 @@ class LogsBackend(BaseBackend):
         ]
         groups = sorted(groups, key=lambda x: x["logGroupName"])
 
-        index_start = 0
-        if next_token:
-            try:
-                index_start = (
-                    next(
-                        index
-                        for (index, d) in enumerate(groups)
-                        if d["logGroupName"] == next_token
-                    )
-                    + 1
-                )
-            except StopIteration:
-                index_start = 0
-                # AWS returns an empty list if it receives an invalid token.
-                groups = []
-
-        index_end = index_start + limit
-        if index_end > len(groups):
-            index_end = len(groups)
-
-        groups_page = groups[index_start:index_end]
-
-        next_token = None
-        if groups_page and index_end < len(groups):
-            next_token = groups_page[-1]["logGroupName"]
-
-        return groups_page, next_token
+        return groups
 
     def create_log_stream(self, log_group_name, log_stream_name):
         if log_group_name not in self.groups:
@@ -546,12 +517,12 @@ class LogsBackend(BaseBackend):
             raise ResourceNotFoundException()
         log_group = self.groups[log_group_name]
         return log_group.describe_log_streams(
-            descending,
-            limit,
-            log_group_name,
-            log_stream_name_prefix,
-            next_token,
-            order_by,
+            descending=descending,
+            limit=limit,
+            log_group_name=log_group_name,
+            log_stream_name_prefix=log_stream_name_prefix,
+            next_token=next_token,
+            order_by=order_by,
         )
 
     def put_log_events(
